feature: add support for multi region param
