+++ src/commands/code/status.ts
@@ -1,7 +1,6 @@
 import { Flags } from '@oclif/core';
 import { StatusCodes } from 'http-status-codes';
 
-import { addRegionToFlags } from 'commands/utils/region';
 import { AuthenticatedCommand } from 'commands-base/authenticated-command';
 import { APP_VERSION_ID_TO_ENTER, VAR_UNKNOWN } from 'consts/messages';
 import { DynamicChoicesService } from 'services/dynamic-choices-service';
@@ -10,7 +9,9 @@ import { getMondayCodeBuild } from 'src/services/app-builds-service';
 import { HttpError } from 'types/errors';
 import { AppVersionDeploymentStatus } from 'types/services/push-service';
 import logger from 'utils/logger';
-import { getRegionFromString } from 'utils/region';
+import { addRegionToFlags, getRegionFromString } from 'utils/region';
+
+const DEBUG_TAG = 'code_status';
 
 const printDeploymentStatus = (
   appVersionId: number,
@@ -67,6 +68,7 @@ export default class Status extends AuthenticatedCommand {
 
       printDeploymentStatus(appVersionId, deploymentStatus);
     } catch (error: unknown) {
+      logger.debug({ res: error }, DEBUG_TAG);
       if (error instanceof HttpError && error.code === StatusCodes.NOT_FOUND) {
         logger.error(`No deployment found for provided app version id - "${appVersionId || VAR_UNKNOWN}"`);
       } else {
